fix:added buffer region fraction
When injecting particles, a "buffer region fraction" allows us to more
easily tweak the "ease" of injection. The larger the "buffer fraction"
is, the more shifted into the simulation domain the particles are.

Also, a bugfix for the N_physical which should NOT be an integer. This
causes issues in the case of the ions, which have such a short injection
zone due to slower thermal speeds, where the f_over_g --> 0 due to the
rounding down of the N_physical.