feat:swapped unit normalization
The velocity is now normalized by the thermal velocity based on the
input temperature. This is more consistent with the use of this PBJ
code, though a potential future feature can be to swap models and unit
normalization for electromagnetic simulation where it might be more
useful to normalize by the speed of light, c.