feat:modified injection scheme
This now injects particles with velocities selected from a thermal
velocity (maxwellian) distribution at a position that is calculated by
randomizing the distance traveled in one time-step for the input thermal
velocity. The number of simulation markers used is determined by the
particle per cell input (though there is a minimum of one cell used and
so ppc likely determines the number of particles injected as well).