feat:SLPIC effects added
In order to add SLPIC effects, there are some modifications necessary
for the particle position update, velocity update, and the density
calculation. This is defined in Appendix A in the SLPIC paper (2019).

In addition, in the actual code, a beta function which acts on the
velocities in the vx,vy,vz variables must be defined. This function is
then passed to the corresponding routines listed above.

An important subtlety in the calculation of the initial f/g for the
particles comes from (A1c) in the Appendix A of the SLPIC paper. Note
that the SLPIC density is actually a reduction of the velocity space
with the weight being (w_p * beta) so that the original definition of
the particl weight (f/g) in the code must be divided by beta to be
properly calculated and equivalent to before.

At the moment, a "return_one" function is defined as the default
function when no function is passed in. In reality, a code which does
not want to use SLPIC should call the normal functions instead.

A final note on this SLPIC implementation is that the authors of SLPIC
used a "modified midpoint method" for the velocity update due to the
dependence of beta on velocity. In my implementation, the beta is
calculated at the average between the velocities at n-1/2 and n+1/2
using a predictor-corrector approach where the prediction is made by
using the old velocity to calculate the expected velocity at n. This
allows for keep the same Boris push function as the main update with
a SLPIC wrapper to modify the electric field used in the push.